New functionality: Read from file containing meta.
Now, instead of setting all parameters every time
file is to be used with a new project. Instead,
all information is defined in a metafile called
`meta.dat` and located in project root (where the
`Make.py` script is located. The metafile has a
certain syntax that has to be followed but all
information can be defined there.

Will write documentation on the compile script and
the meta file later when time.